backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
pydantic==2.10.4
httpx>=0.24.0
beautifulsoup4>=4.12.0
weasyprint==63.1
markdown==3.7
